rename NotebookGitRepo to NotebookRepoWithVersionControl
